requirements-api.txt
# VoiceTransl API Dependencies
# Core API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client for URL downloads
httpx>=0.25.0
requests>=2.31.0

# File processing and validation
python-multipart>=0.0.6
python-magic>=0.4.27
mutagen>=1.47.0

# Rate limiting and monitoring
slowapi>=0.1.9
psutil>=5.9.0

# Async support
asyncio-throttle>=1.0.2

# Logging and monitoring
structlog>=23.2.0

# Configuration management
PyYAML>=6.0.1

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools (optional)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Existing VoiceTransl dependencies (ensure compatibility)
# Note: These should already be installed with the main application
# torch>=2.0.0
# transformers>=4.30.0
# whisper>=1.1.10
# openai-whisper>=20231117
# yt-dlp>=2023.11.16
# ffmpeg-python>=0.2.0
